+++ src/init.ts
@@ -176,13 +176,17 @@ Hooks.once('ready', async () => {
 
   setTimeout(() => {
     overlay = new EPOverlay();
-    SlWindow.container = overlay;
     document.body.append(overlay);
+    // SlWindow.container = overlay;
+    const windowContainer = document.createElement('div');
+    windowContainer.className = 'ep-window-container';
+    SlWindow.container = windowContainer
+    document.body.append(windowContainer);
+   
 
     tooltip = document.createElement('sl-tooltip');
-    tooltip.slot = 'tooltip';
+    // tooltip.slot = 'tooltip';
     overlay.append(tooltip);
-    // document.body.append(tooltip);
     document.body.classList.add('ready');
     Hooks.call('ep-ready', true);
     // const sceneView = document.createElement('scene-view');